PIP dependency support (#1)
* Check in an initial version of the pip rules.

This is broken off of my prototype repository, and there are a handful of TODOs left to resolve.

* Remove some vestigial convenience code.

Elaborate on why the TODO to rewrite {pip,whl}.sh as Python is harder than it looks.

* Incorporate Code Review Feedback

* First cut at skydoc-based documentation.

* Move whl.sh into a Python script that has unit testing.

* Migrate pip.sh into piptool.py

This migrates the logic from pip.sh into piptool.py, which should improve portability by removing the bash dependency.

This also has the beginnings of wrapping piptool as a closed redistributable that doesn't rely on a system-installed copy of PIP, but instead uses these rules to pull pip into a PAR bundle.  Besides needing to work out the details of releasing and redistributing the PAR, we have two unresolved issues:
 * When bundled as a PAR (vs. py_binary), piptool seems to pick up the system-installed version of pip.
 * When bundled as a PAR, piptool sometimes sees cert issues resolving requirements (similar to what we see with httplib2).

* Address the cert issue in piptool as a PAR.

With this change I am able to build/test my PR on my macbook without pip installed.  The only additional change I have locally is to switch from running piptool.py as a simple .py file to downloading/using a PAR built from this change.

I believe we still have the problem that the .par picks the host's version of pip instead of our embedded copy, but I haven't reverified that issue still exists (this just does nothing to address that issue).

* Fix assorted buildifier issues.

* Fix a typo in docs

* Incorporate @duggelz code review feedback.

* Incorporate review feedback

move python tools under a top-level rules_python package

simplify version_test

* Adopt a canonical naming format for imported whl_library rules.

`whl_library` rules generated by `pip_import` are now named as:
```
   pypi__{distribution}_{version}
```

Substituting illegal characters (e.g. `-`, `.`) with underscores.

* Move the piptool dependency into a .PAR file.

Add a script for updating the tools and document this and the docs script.

* Fix buildifier issues.