Make pip_import more verbose
Don't supress output from piptool.par.  This allows users to see
what bazel is doing when it appears stuck while downloading wheels.

While the output is a bit on the spammy side now, the common case
where the pips are already up-to-date will be completely silent
(of course).  Also, the more pressing problem is that people will
just kill bazel/dazel when they think it's stuck - being too spammy
is much less of a real problem.