Support referencing wheels via absolute URLs
pip requirements.txt format allows specifying requirements via
absolute network URLs, in addition to package==version syntax.
If that syntax is detected in requirements.txt, we pass the url
along to the bazel rules, so that the wheel will be pulled from
the URL instead of using "pip wheel package==version" command.