Add pip_version_proxy to simplify picking correct pip deps
The new pip_version_proxy rule generates a requirements.bzl file that provides
a requirements() function that resolves to the correct version
of python dependencies depending on the current python runtime.
This is used via the WORKSPACE pattern:

    pip_import(
        name = "pip_deps2",
        requirements = ["//:requirements-pip.txt"],
        python = "@python2//:bin/python",
    )
    pip_import(
        name = "pip_deps3",
        requirements = ["//:requirements-pip.txt"],
        python = "@python3//:bin/python",
    )
    load("@pip_deps2//:requirements.bzl", pip_install2 = "pip_install")
    pip_install2()
    load("@pip_deps3//:requirements.bzl", pip_install3 = "pip_install")
    pip_install3()

    pip_version_proxy(
        name = "pip_deps",
        define = "python",
        values = {
            "py2": "@pip_deps2",
            "py3": "@pip_deps3",
        },
    )

With this code, requirement("foo") would map to @pip_deps2 when running
"bazel build --define python=py2", and to @pip_deps3 when running with
"bazel build --define python=py3".