Make wheel build deterministic
Use a fixed --build dir when invoking "pip wheel", to make debug symbols
stable.

Pass various defines like -D__DATE__=redacted to the wheel build to avoid
getting timestamps into the binaries.

If the wheel is built locally, strip timestamps from the whl zip file.

Upgrade wheel package used by piptool to a later version (old one would
produce non-deterministic wheels on python3).

Support "sha256" attribute in the "download_or_build_wheel" rule, and use
it in repository_ctx.download (when downloading) or pass to "piptool build",
where we will check that the built wheel digest matches the expected one.
(Note that "piptool resolve" does not yet emit the "sha256" attribute to
the checked-in requirements.bzl, that will be done in a follow-up change).

The build can still be non-deterministic when executed on different
machines, since it's not hermetic.  But when built inside a container,
it should be.