Support adding new wheels via requirements_overrides
Sometimes it's convenient to add a new wheel via requirements_overrides,
to add a user-built / self-hosted wheel, or to work around wheels that
cannot be resolved.  As an example of the latter case, add nvidia-ml-py
which must be installed with "pip install nvidia-ml-py==375.53", but
actually produces a wheel with version 375.53.1.

Issue #19